CLOUDERA-BUILD : CDH-23366 : PIG ILLUSTRATE command broken
getScriptHash() needs logical plan set in the PigContext's execution engine,
which is not true for ILLUSTRATE.

In fact, in ILLUSTRATE's case, the logical plan is never set for PigContext
execution engine. I'm not entirely sure why, but there is bunch of
"if(!pigContext.inIllustrator)" code in Pig, so ILLUSTRATE as a special case is
apparently pretty common :(

This is related to a CDH-only Navigator change, so I can't really compare with
upstream -- where much of the related code has been refactored anyways.

Change-Id: I072560c3da6224930ee1ca93d11f5eb46fb727db